+++ app/src/api/types/incoming/control.ts
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Timestamp } from '@opentalk/common';
 import {
   BackendParticipant,
   ErrorStruct,
@@ -10,6 +9,7 @@ import {
   ParticipantId,
   ParticipantMediaState,
   Tariff,
+  Timestamp,
 } from '@opentalk/common';
 import { InitialAutomod } from '@opentalk/components';