+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Notify moderator if participant limit has been reached (!877)
 - Show room blocker for a participant, that tries to join a full room (!876)
 - Add time_limit quotas in conference (!875)
+- Deactivate modules from room tariff info (!873)
 
 ## 1.2.0